refactor: move refresh button to client
